+++ src/common/labels.rs
@@ -50,17 +50,29 @@ impl LabelElem {
 
     #[allow(dead_code)]
     pub fn new_def(name: smartstring::alias::String, definition: i128) -> LabelElem {
-        let mut elem = LabelElem::new();
-        elem.set_name(name);
-        elem.set_def(definition);
-        elem
+        let ltype = LabelType::Uninit;
+        let referenced = false;
+
+        LabelElem {
+            scope: !name.starts_with('.'),
+            name,
+            definition,
+            ltype,
+            referenced 
+        }
     }
 
     pub fn new_refd(name: smartstring::alias::String) -> LabelElem {
-        let mut elem = LabelElem::new();
-        elem.set_name(name);
-        elem.set_refd();
-        elem
+        let definition: i128 = 0;
+        let ltype = LabelType::Uninit;
+
+        LabelElem {
+            scope: !name.starts_with('.'), 
+            name,
+            definition,
+            ltype,
+            referenced: true
+        }
     }
 
     pub fn combine(&mut self, other: &LabelElem) -> Result<&str, CommonError> {
@@ -194,7 +206,7 @@ impl LabelRecog {
             Some(label) => label.set_refd(),
             None => {
                 let mut label = LabelElem::new_refd(label_str.clone());
-                label.set_scope(true);
+                label.set_scope(!label_str.starts_with('.'));
                 let _ = self.insert_label(label);
             },
         }